[feat] add configuration space read function
