+++ support/devices/virtio/src/lib.rs
@@ -297,42 +297,3 @@ impl VirtioMmio {
         self.get_virtio_mmio(VIRTIO_MMIO_QUEUE_SIZE_MAX) != 0
     }
 }
-
-// Virtio MMIO によって設定されたQueue情報
-#[repr(C)]
-#[derive(Debug)]
-pub struct VirtioMmioRegister {
-    pub queue_size: u32,
-    pub queue_sel: u32,
-    pub queue_ready: u32,
-    pub desc_address: u64,
-    pub driver_address: u64,
-    pub device_address: u64,
-    pub status: u32,
-    pub device_features_low: u32,
-    pub device_features_high: u32,
-    pub device_features_sel: u32,
-    pub driver_features_low: u32,
-    pub driver_features_high: u32,
-    pub driver_features_sel: u32,
-}
-
-impl VirtioMmioRegister {
-    pub const fn new() -> Self {
-        VirtioMmioRegister {
-            queue_size: 0,
-            queue_sel: 0,
-            queue_ready: 0,
-            desc_address: 0,
-            driver_address: 0,
-            device_address: 0,
-            status: 0,
-            device_features_low: 0,
-            device_features_high: 1, // A driver MUST accept VIRTIO_F_VERSION_1 if it is offered.
-            device_features_sel: 0,
-            driver_features_low: 0,
-            driver_features_high: 0,
-            driver_features_sel: 0,
-        }
-    }
-}
